+++ migrations/20220830091135_add_average_nb_of_days_to_user_details.ts
@@ -1,7 +1,7 @@
 
 export async function up(knex) {
     return knex.schema.table('user_details', (table) => {
-        table.integer('average_nb_of_days')
+        table.float('average_nb_of_days')
     })
 }
